Game poster, cover, trending image update, video delete, screenshot delete option done

+++ app/Repositories/Admin/GameRepository.php
@@ -13,6 +13,7 @@ use App\Models\Screenshots;
 use App\Models\Video;
 use App\Models\VideoUrl;
 use Illuminate\Support\Str;
+use Illuminate\Support\Facades\File;
 
 class GameRepository
 {
@@ -64,6 +65,7 @@ class GameRepository
         $game_data['slug'] = Str::slug($game_data['name']);
         $game_data['publisher'] = $game_data['publisher'] ?? 'Testing publisher';
         $game_data['description'] = $game_data['description'] ? $game_data['description'] : 'Testing description';
+
         if ($request->hasFile('trending_url')) {
             $trending = $request->file('trending_url');
             $trending_name = $game_data['name'] . '-trending-' . auth()->user()->id . '-' . time() . $trending->getClientOriginalName();
@@ -71,6 +73,7 @@ class GameRepository
             $trending->storeAs('game-image', $trending_name);
             $game_data['trending_url'] = 'storage/' . $path;
         }
+
         if ($request->hasFile('cover_url')) {
             $cover = $request->file('cover_url');
             $cover_name = $game_data['name'] . '-cover-' . auth()->user()->id . '-' . time() . $cover->getClientOriginalName();
@@ -78,6 +81,7 @@ class GameRepository
             $cover->storeAs('game-image', $cover_name);
             $game_data['cover_url'] = 'storage/' . $path;
         }
+
         if ($request->hasFile('poster_url')) {
             $poster = $request->file('poster_url');
             $poster_name = $game_data['name'] . '-poster-' . auth()->user()->id . '-' . time() . $poster->getClientOriginalName();
@@ -85,7 +89,7 @@ class GameRepository
             $poster->storeAs('game-image', $poster_name);
             $game_data['poster_url'] = 'storage/' . $path;
         }
-//        return $game_data;
+
         $game = Game::create($game_data);
 
         $game->genres()->sync($request->genres, false);
@@ -140,7 +144,6 @@ class GameRepository
      */
     public function update($request, $id)
     {
-
         $game = Game::findOrFail($id);
         $data = $request->only(['name', 'released', 'rating', 'description', 'is_trending', 'publisher', 'developer']);
 
@@ -165,6 +168,34 @@ class GameRepository
         if (isset($data['is_trending'])) {
             $game->is_trending = $data['is_trending'];
         }
+
+        if ($request->hasFile('trending_url')) {
+            $trending = $request->file('trending_url');
+            $trending_name = $data['name'] . '-trending-' . auth()->user()->id . '-' . time() . $trending->getClientOriginalName();
+            $path = "game-image/" . $trending_name;
+            $trending->storeAs('game-image', $trending_name);
+            deleteFile([$game->trending_url]);
+            $game->trending_url = 'storage/' . $path;
+        }
+
+        if ($request->hasFile('cover_url')) {
+            $cover = $request->file('cover_url');
+            $cover_name = $data['name'] . '-cover-' . auth()->user()->id . '-' . time() . $cover->getClientOriginalName();
+            $path = "game-image/" . $cover_name;
+            $cover->storeAs('game-image', $cover_name);
+            deleteFile([$game->cover_url]);
+            $game->cover_url = 'storage/' . $path;
+        }
+
+        if ($request->hasFile('poster_url')) {
+            $poster = $request->file('poster_url');
+            $poster_name = $data['name'] . '-poster-' . auth()->user()->id . '-' . time() . $poster->getClientOriginalName();
+            $path = "game-image/" . $poster_name;
+            $poster->storeAs('game-image', $poster_name);
+            deleteFile([$game->poster_url]);
+            $game->poster_url = 'storage/' . $path;
+        }
+
         $game->save();
 
         $game->genres()->sync($request->genres);
@@ -186,8 +217,11 @@ class GameRepository
             }
         }
 
-        if ($request->has('video_url')) {
+        if ($request->get('video_url')) {
             foreach ($request->video_url as $value) {
+                if (empty($value)) {
+                    continue;
+                }
                 $video_name = $game->name . '-Video-' . auth()->user()->id . '-' . time();
                 $video_data = $request->only(['game_id', 'name', 'url']);
                 $video_data['game_id'] = $game->id;
@@ -206,7 +240,9 @@ class GameRepository
 
             if (isset($game->assets[0]) && $game->assets[0] != null) {
                 $oldImg = $game->assets[0]->url;
-                unlink($oldImg);
+                if (file_exists($oldImg)) {
+                    unlink($oldImg);
+                }
 
                 $img = Asset::find($game->assets[0]->id);
                 $img->name = $image_name;
@@ -254,9 +290,11 @@ class GameRepository
     {
         $screenshots = Screenshots::findOrFail($id);
         if ($screenshots) {
+            deleteFile([$screenshots->url]);
             $screenshots->delete();
             return true;
         }
+
         return false;
     }
